Reorganised AVX related code and separate them out into one file.
