+++ percival_parallel/percival_ADC_decode_pipe.cpp
@@ -9,32 +9,8 @@
 #include "percival_processing.h"
 #include "percival_parallel.h"
 #include "percival_data_validity_checks.h"
+#include "percival_avx.h"
 #include "tbb/pipeline.h"
-
-void percival_ADC_decode_pipe(
-		const percival_frame<unsigned short int> & input,
-		percival_frame<float> & output,
-		const percival_calib_params & calib_params,
-		percival_frame<unsigned short int> gain,
-		percival_frame<unsigned short int> fine,
-		percival_frame<unsigned short int> coarse,
-		percival_frame<float> calibrated,
-		bool store_gain)
-{
-	unsigned int height = input.height;
-	unsigned int width = input.width;
-
-	//	percival_frame<unsigned short int> gain(height, width);
-	//	percival_frame<unsigned short int> fine(height, width);
-	//    percival_frame<unsigned short int> coarse(height, width);
-	//	percival_frame<float> calibrated(height, width);
-	unsigned int NoOfPixels = output.width * output.height;
-
-	percival_ADC_decode_pipe_p< tbb::blocked_range<unsigned int> > percival_ADC_decode_pipe_p (input , output, calib_params, gain, fine, coarse, calibrated);
-	tbb::parallel_for( tbb::blocked_range<unsigned int>(0, NoOfPixels), percival_ADC_decode_pipe_p, tbb::auto_partitioner());
-
-}
-
 
 /*===============================================================================================================================*/
 /*
@@ -104,6 +80,39 @@ public:
 	}
 };
 
+class ADC_decode_combined_filter : public tbb::filter{
+private:
+	const unsigned int grain_size;		/* size of loop */
+	percival_calib_params calib_params;
+
+	percival_algorithm_avx< percival_range_iterator_mock_p > algorithm;
+	percival_range_iterator_mock_p range;
+
+public:
+	ADC_decode_combined_filter(
+			const percival_frame<unsigned short> & sample,
+			const percival_frame<unsigned short> & reset,
+			percival_frame<float> & output,
+			const percival_calib_params & calib_params,
+			const unsigned int grain_size):
+				tbb::filter(/*is_serial=*/false),
+				calib_params(calib_params),
+				grain_size(grain_size),
+				algorithm(sample, reset, output, calib_params),
+				range(0,1){}
+
+
+	void* operator()(void* input){
+		unsigned int * offset_ptr = static_cast < unsigned int* >(input);
+		unsigned int offset = *offset_ptr;
+		/* range to loop over */
+		range.lower = offset;
+		range.upper = grain_size + offset;
+		/*running the algorithm*/
+		algorithm(range);
+		return NULL;	/*return to next stage if needed*/
+	}
+};
 
 void percival_ADC_decode_combined_pipeline(
 		const percival_frame<unsigned short int> & sample,
@@ -143,11 +152,6 @@ void percival_ADC_decode_combined_pipeline(
 	unsigned int *offset_ptr = & offset_arr[0];
 
 	/* Initialising the input image struct */
-	head_to_CDS CDS_input;
-	CDS_input.input_sample = sample;
-	CDS_input.input_reset = reset;
-	CDS_input.output = output;
-
 	/*
 	 *  Constructing the pipeline
 	 *
@@ -156,24 +160,12 @@ void percival_ADC_decode_combined_pipeline(
 	percival_pipeline_stream_generator Input(offset_ptr, grain_size, NoOfPixels, max_tokens);
 	pipeline.add_filter( Input );
 
-	ADC_decode_combined_filter<head_to_CDS> ADC_decode_CDS ( CDS_input, calib_params, grain_size );
+	ADC_decode_combined_filter ADC_decode_CDS ( sample, reset, output, calib_params, grain_size );
 	pipeline.add_filter( ADC_decode_CDS );
 
 	pipeline.run( max_tokens );
 
 	pipeline.clear();
-
-	/*
-	 *
-	 *  Parallel_for version for comparison purposes
-	 *
-	 */
-
-//	unsigned int NoOfPixels = sample.width * sample.height;
-//
-//	percival_algorithm_p< head_to_CDS, tbb::blocked_range<unsigned int> > percival_p (CDS_input,calib_params);
-//	tbb::parallel_for( tbb::blocked_range<unsigned int>(0, NoOfPixels, grain_size), percival_p, tbb::auto_partitioner());
-
 }
