+++ percival_parallel/percival_parallel.h
@@ -13,6 +13,7 @@
 #include "tbb/parallel_for.h"
 #include "percival_processing.h"
 #include "percival_functors.h"
+#include "percival_avx.h"
 
 /* Use the following to set optimised parameters
 struct optimised_parameter{
@@ -21,12 +22,6 @@ struct optimised_parameter{
 	const static unsigned int max_threads;
 };
 */
-
-void percival_unit_ADC_decode_pf(
-		const percival_frame<unsigned short int> & input,
-		percival_frame<unsigned short int> & Coarse,
-		percival_frame<unsigned short int> & Fine,
-		percival_frame<unsigned short int> & Gain);
 
 void percival_unit_ADC_calibration_pf(
 		const percival_frame<unsigned short int> & Coarse,
@@ -52,39 +47,8 @@ void percival_ADC_decode_combined_pipeline(
 		unsigned int grain_size = 3528,
 		unsigned int max_tokens = 20);
 
-template<typename input_type>
-class ADC_decode_combined_filter : public tbb::filter{
-private:
-	const unsigned int grain_size;		/* size of loop */
-	percival_calib_params calib_params;
-
-	percival_algorithm_avx< input_type, percival_range_iterator_mock_p > algorithm;
-	percival_range_iterator_mock_p range;
-
-public:
-	ADC_decode_combined_filter(
-			input_type & input,
-			const percival_calib_params & calib_params,
-			const unsigned int grain_size):
-				tbb::filter(/*is_serial=*/false),
-				calib_params(calib_params),
-				grain_size(grain_size),
-				algorithm(input, calib_params),
-				range(0,1){}
-
-
-	void* operator()(void* input){
-		unsigned int * offset_ptr = static_cast < unsigned int* >(input);
-		unsigned int offset = *offset_ptr;
-		/* range to loop over */
-		range.lower = offset;
-		range.upper = grain_size + offset;
-		/*running the algorithm*/
-		algorithm(range);
-		return NULL;	/*return to next stage if needed*/
-	}
-};
-
+class percival_pipeline_stream_generator;
+class ADC_decode_combined_filter;
 
 
 #endif /* PERCIVAL_PARALLEL_PERCIVAL_PARALLEL_H_ */
